% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_emissions.R
\name{add_emissions}
\alias{add_emissions}
\title{Add emissions parameters to a dispersion model}
\usage{
add_emissions(model, rate = NULL, duration = NULL, start_day = NULL,
  start_hour = NULL, name = NULL)
}
\arguments{
\item{model}{a SplitR modeling object.}

\item{rate}{the rate of emissions for the pollutant
in mass units per hour.}

\item{duration}{the duration of emissions in hours.}

\item{start_day}{the day that the emissions will 
begin. This should take the form of a
single-length vector for a day (\code{"YYYY-MM-DD"}).}

\item{start_hour}{a single daily hour as an
integer hour (from \code{0} to \code{23}).}

\item{name}{an identifier for the group of emissions
parameters.}
}
\description{
Add a set of emissions parameters to
a dispersion model object. Multiple sets of
emissions parameters can can be added to a single
dispersion model object.
}
