% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_params.R
\name{add_params}
\alias{add_params}
\title{Add model parameters}
\usage{
add_params(model, lat = NULL, lon = NULL, height = NULL,
  duration = NULL, run_type = NULL, run_day = NULL, run_range = NULL,
  run_years = NULL, daily_hours = NULL, backtrajectory = NULL,
  met_type = NULL, vert_motion = NULL, model_height = NULL)
}
\arguments{
\item{model}{a SplitR modeling object}
}
\description{
Add modelling parameters to a model
object
}

