% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_source.R
\name{add_source}
\alias{add_source}
\title{Add source parameters to a dispersion model}
\usage{
add_source(model, name = NULL, lat = NULL, lon = NULL,
  height = NULL, release_start = NULL, release_end = NULL,
  rate = NULL, pdiam = NULL, density = NULL, shape_factor = NULL,
  ddep_vel = NULL, ddep_mw = NULL, ddep_a_ratio = NULL,
  ddep_d_ratio = NULL, ddep_hl_coeff = NULL, wdep_hl_coeff = NULL,
  wdep_in_cloud = NULL, wdep_below_cloud = NULL, rad_decay = NULL,
  resuspension = NULL)
}
\arguments{
\item{model}{A dispersion model object.}

\item{name}{An identifier for the source}

\item{lat, lon, height}{The source position in terms of latitude and longitude
(both in decimal degrees), and height in meters above ground level.}

\item{release_start, release_end}{The beginning and ending times for the
source emissions.}

\item{rate}{The particle emission rate.}

\item{pdiam}{The particle diameter in units of micrometers.}

\item{density}{The particle density in units of g/cm3.}

\item{shape_factor}{The particle shape factor as a value in the range
\code{0}-\code{1}.}

\item{ddep_vel}{The dry deposition velocity in units of m/s.}

\item{ddep_mw}{The molecular weight in units of g/mol.}

\item{ddep_a_ratio}{The dry deposition A ratio.}

\item{ddep_d_ratio}{The dry deposition D ratio.}

\item{ddep_hl_coeff}{The Henry's Law value associated with dry deposition.}

\item{wdep_hl_coeff}{The Henry's Law value associated with wet deposition.}

\item{wdep_in_cloud}{The in-cloud wet deposition rate in units of L/L.}

\item{wdep_below_cloud}{The below cloud wet deposition rate in units of 1/s.}

\item{rad_decay}{The rate of radioactive decay, in units of days.}

\item{resuspension}{The pollutant resuspension factor in units of 1/m.}
}
\description{
Add a set of source parameters to a dispersion model.
}
