% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_disp_model.R
\name{create_disp_model}
\alias{create_disp_model}
\title{Create a dispersion model}
\usage{
create_disp_model(name = NULL)
}
\arguments{
\item{name}{an optional name for the dispersion
model object.}
}
\description{
Create a dispersion model object to
begin a modelling pipeline.
}
