% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dispersion_model.R
\name{create_dispersion_model}
\alias{create_dispersion_model}
\title{Create a dispersion model}
\usage{
create_dispersion_model()
}
\description{
Create a dispersion model object to begin a modeling pipeline.
}
