% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traj_model.R
\name{create_traj_model}
\alias{create_traj_model}
\title{Create a trajectory model}
\usage{
create_traj_model(name = NULL)
}
\arguments{
\item{name}{an optional name for the trajectory
model object.}
}
\description{
Create a trajectory model object to
begin a modelling pipeline.
}
