% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trajectory_model.R
\name{create_trajectory_model}
\alias{create_trajectory_model}
\title{Create a trajectory model}
\usage{
create_trajectory_model(traj_name = NULL)
}
\description{
Create a trajectory model object to
begin a modelling pipeline.
}

