\name{Dispersion model demonstration}
\alias{demo.dispersion}
\title{
Run an interactive demonstration for a dispersion model
}
\description{
Demonstrates the steps required to set up a project and get dispersion models running.
}
\usage{
demo.dispersion()
}
\author{
Richard Iannone
}
\examples{
# Run the dispersion demo
demo.dispersion()
}
