\name{Trajectory model demonstration}
\alias{demo.trajectory}
\title{
Run an interactive demonstration for a trajectory model
}
\description{
Demonstrates the steps required to set up a project and get trajectory models running.
}
\usage{
demo.trajectory()
}
\author{
Richard Iannone
}
\examples{
# Run the trajectory demo
demo.trajectory()
}
