% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disp.stats.particle_heights}
\alias{disp.stats.particle_heights}
\title{Get statistics for dispersed particle heights}
\usage{
disp.stats.particle_heights(dispersion_df, stats)
}
\arguments{
\item{dispersion_df}{a data frame generated from dispersion output data returned from the 'dispersion.read' function.}

\item{stats}{the type of statistical analysis to perform. Currently, "fivenum" is available.}
}
\description{
Get statistics for dispersed particle heights
}
\examples{
# Get a five-number statistical summary of particle heights at every hour
disp.stats.particle_heights(dispersion_df = disp_output,
                            stats = "fivenum")
}

