% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disp_stats_particle_heights}
\alias{disp_stats_particle_heights}
\title{Get statistics for dispersed particle heights}
\usage{
disp_stats_particle_heights(dispersion_df, stats)
}
\arguments{
\item{dispersion_df}{a data frame generated from dispersion output data returned from the 'dispersion.read' function.}

\item{stats}{the type of statistical analysis to perform. Currently, "fivenum" is available.}
}
\description{
This function computes statistics focussed on the heights of particles at every time interval.
}
\examples{
\dontrun{
# Get a five-number statistical summary of particle heights at every hour
disp_stats_particle_heights(dispersion_df = disp_output,
                            stats = "fivenum")
}
}

