% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_stats_particle_through_region.R
\name{disp_stats_particle_through_region}
\alias{disp_stats_particle_through_region}
\title{Get statistics for dispersed particles travelling through a specified region}
\usage{
disp_stats_particle_through_region(dispersion_df, stats, lat, lon, heights)
}
\arguments{
\item{dispersion_df}{a data frame generated from dispersion output data returned from the 'dispersion_read' function.}

\item{stats}{the type of statistical analysis to perform. Currently, "frequencies" is available.}

\item{lat}{a numeric vector containing latitude bounds}

\item{lon}{a numeric vector containing longitude bounds}

\item{heights}{a numeric vector containing height bounds}
}
\description{
This function computes statistics focussed on particles travelling through a specified volume of air.
}
\examples{
\dontrun{
# Get a statistical summary of particles available in region every hour
disp_stats_particle_through_region(dispersion_df = disp_output,
                                   stats = "frequencies",
                                   lat = c(49.05, 49.50),
                                   lon = c(-127.2, -128.0),
                                   heights = c(0, 1000))
}
}

