\name{Remove dispersion preset information}
\alias{dispersion.preset.delete}
\title{
Remove a preset for HYSPLIT dispersion model runs
}
\description{

}
\usage{
dispersion.preset.delete(read = NULL, numbers = NULL, interactive = TRUE)
}
\arguments{
  \item{read}{

}
  \item{numbers}{

}
  \item{interactive}{

}
}
\author{
Richard Iannone
}
\examples{

}
