% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion.preset.list}
\alias{dispersion.preset.list}
\title{Get list of presets}
\usage{
dispersion.preset.list(read = NULL, search = NULL,
  path_wd = "~/Documents/SplitR/Working/")
}
\arguments{
\item{read}{the type of preset to get. The three preset types are "emissions", "grids", and "species". Only a single type can be specified.}

\item{search}{a search term corresponding to the name of a preset of the type provided within the argument 'read'. If an exact match is found (case-sensitive), then the number corresponding to the matched preset is provided.}

\item{path_wd}{}
}
\description{
Get list of presets
}
\examples{
# Get a list of 'species' presets
dispersion.preset.list(read = "species")

# Find the number for the 'emissions' preset named 'E134'
dispersion.preset.list(read = "emissions", search = "E134")
}

