% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion.read}
\alias{dispersion.read}
\title{Read HYSPLIT dispersion output files}
\usage{
dispersion.read(path_output_files)
}
\arguments{
\item{path_output_files}{}
}
\description{
Read HYSPLIT dispersion output files
}
\examples{
# Process all dispersion output files in the specified output directory
dispersion.df <- dispersion.read(path_output_files =
                                 "~/hysplit/trajectory")
}

