% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_plot.R
\name{dispersion_plot}
\alias{dispersion_plot}
\title{Plot HYSPLIT dispersion model output onto a map}
\usage{
dispersion_plot(disp_df, color_scheme = "cycle_hues")
}
\arguments{
\item{disp_df}{a dispersion data frame, typically
created from use of the \code{hysplit_dispersion}
function with the value for \code{return_disp_df}
set to \code{TRUE}.}

\item{color_scheme}{defines the appearance of
multiple trajectories in a single plot. Current
options are \code{cycle_hues} (the default), and
\code{increasingly_gray}.}
}
\description{
The function plots modeled dispersion
particles onto a map with information provided
at time interval.
}

