% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion_preset_list}
\alias{dispersion_preset_list}
\title{Get list of presets}
\usage{
dispersion_preset_list(read = NULL, search = NULL, path_wd)
}
\arguments{
\item{read}{the type of preset to get. The three preset types are "emissions", "grids", and "species". Only a single type can be specified.}

\item{search}{a search term corresponding to the name of a preset of the type provided within the argument 'read'. If an exact match is found (case-sensitive), then the number corresponding to the matched preset is provided.}

\item{path_wd}{a full path should be provided for the HYSPLIT working directory since presets will normally reside in this folder.}
}
\description{
This function allows you to generate a simple list of presets of a given type. If no type is provided, the user is prompted to enter a type. The list is presented as a series of oneline summaries that are numbered in the order they appear in their respective files. This numbering is important, as the 'hysplit.dispersion' function calls presets by these numbers. It is possible to search for a preset of a certain type by its name through use of the 'search' argument. If there is an exact match by name, then the number associated with that preset is returned.
}
\examples{
\dontrun{
# Get a list of 'species' presets
dispersion_preset_list(read = "species",
                       path_wd = "C:\\\\hysplit4\\\\working\\\\")

# Find the number for the 'emissions' preset named 'E134'
dispersion_preset_list(read = "emissions", search = "E134",
                       path_wd = "~/Documents/Hysplit/working/")
}
}

