% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.edas40}
\alias{get.met.edas40}
\title{Get EDAS40 meteorology data files}
\usage{
get.met.edas40(files = NULL, years = NULL, months = NULL,
  path_met_files = "~/Documents/SplitR/Met/")
}
\arguments{
\item{files}{}

\item{years}{}

\item{months}{}

\item{path_met_files}{}
}
\description{
Get EDAS40 meteorology data files
}
\examples{
# Download a year of EDAS40 data files
get.met.edas40(years = 2012)
}

