% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.forecast.nam}
\alias{get.met.forecast.nam}
\title{Read NAM meteorological forecast data files}
\usage{
get.met.forecast.nam(path_met_files = "~/Documents/SplitR/Met/")
}
\arguments{
\item{path_met_files}{}
}
\description{
This function downloads NAM meteorology forecast data files from the NOAA FTP server and saves them to a specified folder.
}

