% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.gdas0p5}
\alias{get.met.gdas0p5}
\title{Get GDAS0.5 meteorology data files}
\usage{
get.met.gdas0p5(files = NULL, years = NULL, months = NULL, days = NULL,
  path_met_files)
}
\arguments{
\item{files}{}

\item{years}{}

\item{months}{}

\item{days}{}

\item{path_met_files}{}
}
\description{
This function downloads GDAS0.5 meteorology data files from the NOAA FTP server and saves them to a specified folder. Files can be downloaded either by specifying a list of filenames (in the form of "RP[YYYY][MM].gbl") or through bulk download of a year of files.
}

