% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.nam12}
\alias{get.met.nam12}
\title{Get NAM12 meteorology data files}
\usage{
get.met.nam12(files = NULL, years = NULL, months = NULL,
  path_met_files = "~/Documents/SplitR/Met/")
}
\arguments{
\item{files}{}

\item{years}{}

\item{months}{}

\item{path_met_files}{}
}
\description{
Get NAM12 meteorology data files
}

