% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_era5.R
\name{get_met_era5}
\alias{get_met_era5}
\title{Get ERA5 meteorology data files}
\usage{
get_met_era5(days, duration, direction, path_met_files)
}
\arguments{
\item{days}{A vector of days that the model will run. This is combined with
the \code{daily_hours} to produce a series of date-times.}

\item{duration}{The duration of each model run (whether it is in the forward
direction or running backwards) in hours.}

\item{direction}{An option to select whether to conduct the model in the
\code{"forward"} (default) or \code{"backward"} directions.}

\item{path_met_files}{A full directory path to which the meteorological data
files will be saved.}
}
\description{
Generates list of file names for ERA5 dataset. Cannot be
downloaded from arlftp since it is not archived there, must
be downloaded/preprocessed from ECMWF via conversion tool
provided w/ HYSPLIT distro. Comment out download portion,
assume data are present in path_met_files. -RPF 191230.
}
