% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_met_forecast.nam}
\alias{get_met_forecast.nam}
\title{Read NAM meteorological forecast data files}
\usage{
get_met_forecast.nam(path_met_files)
}
\arguments{
\item{path_met_files}{a full path should be provided for the location of the meteorological data files; downloaded files will be saved in this location.}
}
\description{
This function downloads NAM meteorology forecast data files from the NOAA FTP server and saves them to a specified folder.
}

