% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_forecast_nam.R
\name{get_met_forecast_nam}
\alias{get_met_forecast_nam}
\title{Read NAM meteorological forecast data files}
\usage{
get_met_forecast_nam(path_met_files)
}
\arguments{
\item{path_met_files}{a full path should be provided
for the location of the meteorological data files;
downloaded files will be saved in this location.}
}
\description{
This function downloads NAM meteorology
forecast data files from the NOAA FTP server and
saves them to a specified folder.
}
