% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_gdas1.R
\name{get_met_gdas1}
\alias{get_met_gdas1}
\title{Get GDAS1 meteorology data files}
\usage{
get_met_gdas1(files = NULL, path_met_files)
}
\arguments{
\item{files}{A vector list of exact filenames for the meteorological model
files.}

\item{path_met_files}{A full path should be provided for the location of the
meteorological data files. Downloaded files will be saved in this location.}
}
\description{
Downloads GDAS1 meteorology data files from the NOAA FTP server and saves
them to a specified folder. Files can be downloaded by specifying a list of
filenames.
}
