% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_hrrr.R
\name{get_met_hrrr}
\alias{get_met_hrrr}
\title{Get HRRR meteorology data files}
\usage{
get_met_hrrr(files = NULL, path_met_files)
}
\arguments{
\item{files}{a vector list of exact filenames for the HRRR files.}

\item{path_met_files}{a full path should be provided for the location of the
meteorological data files; downloaded files will be saved in this location.}
}
\description{
This function downloads High Resolution Rapid Refresh (HRRR) meteorology data
files from the NOAA FTP server and saves them to a specified folder. Files
can be downloaded only by specifying a list of filenames (in the form of
\code{'RP[YYYY][MM].gbl'}).
}
