% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_reanalysis.R
\name{get_met_reanalysis}
\alias{get_met_reanalysis}
\title{Get reanalysis meteorology data files}
\usage{
get_met_reanalysis(files = NULL, path_met_files)
}
\arguments{
\item{files}{A vector list of exact filenames for the meteorological model
files.}

\item{path_met_files}{A full path should be provided for the location of the
meteorological data files. Downloaded files will be saved in this location.}
}
\description{
Downloads reanalysis meteorology data files from the NOAA FTP server and
saves them to a specified folder. Files can be downloaded by specifying a
list of filenames (in the form of \code{'RP[YYYY][MM].gbl'}).
}
