% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_df.R
\name{get_output_df}
\alias{get_output_df}
\title{Extract model data}
\usage{
get_output_df(model)
}
\arguments{
\item{model}{a SplitR modeling object.}
}
\description{
If a model has been executed, the output
data frame will be available within a SplitR modeling
object and this function will get either the entire
output data frame or a subset of it.
}
