% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_tbl.R
\name{get_output_tbl}
\alias{get_output_tbl}
\title{Extract model data}
\usage{
get_output_tbl(model)
}
\arguments{
\item{model}{A splitr modeling object.}
}
\description{
If a model has been executed, the output tibble will be available within a
splitr modeling object and this function will return the entire output
tibble.
}
