% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traj_df.R
\name{get_traj_df}
\alias{get_traj_df}
\title{Get the trajectory data frame from a model object}
\usage{
get_traj_df(model)
}
\arguments{
\item{model}{a SplitR modeling object}
}
\description{
If a data frame is available within a 
model object, get the entire data frame or a subset
of it
}

