% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hysplit.dispersion.animation}
\alias{hysplit.dispersion.animation}
\title{Create an animated movie from HYSPLIT dispersion model output}
\usage{
hysplit.dispersion.animation(dispersion_df = NULL, start_lat_deg,
  start_long_deg, df_folder_path = NULL, frame_rate = 30,
  movie_output_name = NULL, IM_exec_path = "/opt/local/bin/convert",
  path_output_files = NULL)
}
\arguments{
\item{dispersion_df}{specification of hours of dispersion data to plot.}

\item{start_lat_deg}{the starting latitude (in decimal degrees) for the model run(s).}

\item{start_long_deg}{the starting longitude (in decimal degrees) for the model run(s).}

\item{df_folder_path}{absolute path of the folder containing a dispersion data frame.}

\item{frame_rate}{the desired frame rate of the generated MP4 movie file.}

\item{movie_output_name}{the desired filename for the generate MP4 movie.}

\item{IM_exec_path}{an absolute system path for the ImageMagick 'convert' command.}

\item{path_output_files}{a full path for a location that the dispersion output files were written.}
}
\description{
Create a movie file of dispersed particles from a dispersion data frame
}
\examples{
\dontrun{
# Create an MP4 movie from a data frame returned by the 'hysplit.dispersion' function
hysplit.dispersion.animation(dispersion_df = disp.df,
                             start_lat_deg = 42.83752,
                             start_long_deg = -80.30364,
                             frame_rate = 60,
                             movie_output_name = "dispersion-movie",
                             IM_exec_path = "/opt/local/bin/convert",
                             path_output_files = "~/Documents/SplitR/Output/Movies/"
}
}

