\name{Create configuration file for dispersion runs}
\alias{hysplit.dispersion.config}
\title{
Create or modify the configuration file used for HYSPLIT dispersion models
}
\description{

}
\usage{
hysplit.dispersion.config(create_default_config = TRUE,
                          interactive_mode = FALSE)
}
\arguments{
  \item{create_default_config}{
creates a default SETUP.CFG configuration file and places it in the "/SplitR/Working" folder. This is done in a non-interactive manner so any value for the 'interactive_mode' argument will be disregarded.
}
}
\details{
}
\value{
}
\references{
}
\author{
Richard Iannone
}
\examples{

}
