% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hysplit.dispersion.plot}
\alias{hysplit.dispersion.plot}
\title{Plot HYSPLIT dispersion model output onto maps}
\usage{
hysplit.dispersion.plot(hours = "all", dispersion_df = NULL,
  disp_name = NULL, path_output_files = NULL)
}
\arguments{
\item{hour}{specification of hours of dispersion data to plot.}

\item{dispersion_df}{optionally specify a data frame with dispersion data as an input.}

\item{disp_name}{name of previously stored dispersion project.}

\item{path_output_files}{a full path for a location that the dispersion output files were written.}
}
\description{
The function plots hourly outputs of dispersed particles onto maps.
}

