% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_config_init.R
\name{hysplit_config_init}
\alias{hysplit_config_init}
\title{Create default configuration files for trajectory
and dispersion modelling runs}
\usage{
hysplit_config_init()
}
\description{
This function creates default
\code{SETUP.CFG} and \code{ASCDATA.CFG}
configuration files and places them in the
working directories.
}
\examples{
\dontrun{
# Create a default `SETUP.CFG` and `ASCDATA.CFG`
# files for trajectory and dispersion runs
hysplit_config_init()
}
}

