% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_trajectory.R
\name{hysplit_trajectory}
\alias{hysplit_trajectory}
\title{Conduct HYSPLIT trajectory runs}
\usage{
hysplit_trajectory(lat = 49.263, lon = -123.25, height = 50,
  duration = 24, days = NULL, daily_hours = 0,
  direction = "forward", met_type = "reanalysis", vert_motion = 0,
  model_height = 20000, extended_met = FALSE, traj_name = NULL,
  binary_path = NULL, met_dir = NULL, exec_dir = NULL,
  clean_up = TRUE)
}
\arguments{
\item{lat, lon}{The starting latitude and longitude (in decimal degrees) for
each model run.}

\item{height}{The starting height (in meters above ground level) for each
model run.}

\item{duration}{The duration of each model run (either forward or backward)
in hours.}

\item{days}{A vector of days that the model will run. This is combined with
the \code{daily_hours} to produce a series of date-times.}

\item{daily_hours}{A vector of daily hours for initiations of runs across the
given \code{days}. Use values from from \code{0} to \code{23}.}

\item{direction}{An option to select whether to conduct the model in the
\code{"forward"} (default) or \code{"backward"} directions.}

\item{met_type}{The type of meteorological data files to use. The options
are: \code{"reanalysis"} (NCAR/NCEP global reanalysis data, the default),
\code{"gdas1"} and \code{"gdas0.5"} (Global Data Assimilation System 1-degree and
0.5-degree resolution data), \code{"narr"} (North American Regional Reanalysis),
\code{"gfs0.25"} (Global Forecast System 0.25 degree data), and \code{"nam12"} (North
American Mesoscale Forecast System, 12-km/6-hour resolution data).}

\item{vert_motion}{A numbered option to select the method used to simulation
vertical motion. The methods are: \code{0} (input model data), \code{1} (isobaric),
\code{2} (isentropic), \code{3} (constant density), \code{4} (isosigma), \code{5} (from
divergence), \code{6} (remap MSL to AGL), \code{7} (average data), and \code{8} (damped
magnitude).}

\item{model_height}{The upper limit of the model domain in meters.}

\item{extended_met}{An option to report additional meteorological data along
each output trajectory.}

\item{traj_name}{An optional, descriptive name for the output file
collection.}

\item{binary_path}{An optional path to a HYSPLIT trajectory model binary.
When not specified, the model binary will be chosen from several available
in the package (based on the user's platform).}

\item{met_dir}{An optional file path for storage and access of meteorological
data files.}

\item{exec_dir}{An optional file path for the working directory of the model
input and output files.}

\item{clean_up}{An option to make the \code{exec_dir} directory clean after
completion of all trajectory runs. By default, this is set to \code{TRUE}.}
}
\description{
The function executes single/multiple forward or backward HYSPLIT trajectory
runs using specified meteorological datasets.
}
\examples{
\dontrun{
library(lubridate)

# Run a trajectory model 4 times a day
# for 6 days in 2012 using NCEP/NCAR
# reanalysis data
trajectory <-
  hysplit_trajectory(
    lat = 50.108,
    lon = -122.942,
    height = 100,
    duration = 48,
    days = seq(
      lubridate::ymd("2012-02-22"),
      lubridate::ymd("2012-02-27"),
      by = "1 day"
    ),
    daily_hours = c(0, 6, 12, 18)
  )
}

}
