\name{Archive a SplitR project}
\alias{project.archive}
\title{
Archive a SplitR project.
}
\description{
Shows a numbered list of SplitR projects and the date that each project was created. Any current SplitR project that is currently open will be indicated in this list. Interactive prompts allow the user to archive a project. Note that archiving a project does not delete it from your system, the process effectively moves a project folder to an archive folder and removes the project from the project list.
}
\usage{
project.archive()
}
\author{
Richard Iannone
}
\examples{
# Archive a SplitR project
project.archive()
}
