\name{Show a list of SplitR projects}
\alias{project.list}
\title{
Display a simple list of SplitR projects.
}
\description{
Shows a numbered list of SplitR projects and the date that each project was created.
}
\usage{
project.list()
}
\author{
Richard Iannone
}
\examples{
# Get list of SplitR projects
project.list()
}
