\name{Open an existing SplitR project}
\alias{project.open}
\title{
Display a simple list of SplitR projects and interactively open a project.
}
\description{
Shows a numbered list of SplitR projects and the date that each project was created. Any current SplitR project that is currently open will be indicated in this list. Interactive prompts allow the user to switch to another project.
}
\usage{
project.open()
}
\author{
Richard Iannone
}
\examples{
# Open an existing SplitR project
project.open()
}
