\name{Show a summary of the active SplitR project}
\alias{project.summary}
\title{
Display a summary of actions performed and key settings for the active SplitR project.
}
\description{
Shows a summary of what was recently done, what data is available, what is still required to perform certain model runs and analyse, and some of the key settings for the active SplitR project.
}
\usage{
project.summary()
}
\author{
Richard Iannone
}
\examples{
# Get an information summary of the current SplitR project
project.summary()
}
