% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Create a configuration list for a trajectory or dispersion model}
\usage{
set_config(tratio = 0.75, initd = 0, kpuff = 0, khmax = 9999,
  kmixd = 0, kmix0 = 250, kzmix = 0, kdef = 0, kbls = 1,
  kblt = 2, conage = 48, numpar = 2500, qcycle = 0, efile = NULL,
  tkerd = 0.18, tkern = 0.18, ninit = 1, ndump = 1, ncycl = 1,
  pinpf = "PARINIT", poutf = "PARDUMP", mgmin = 10, kmsl = 0,
  maxpar = 10000, cpack = 1, cmass = 0, dxf = 1, dyf = 1,
  dzf = 0.01, ichem = 0, maxdim = 1, kspl = 1, krnd = 6,
  frhs = 1, frvs = 0.01, frts = 0.1, frhmax = 3, splitf = 1,
  tm_pres = 0, tm_tpot = 0, tm_tamb = 0, tm_rain = 0,
  tm_mixd = 0, tm_relh = 0, tm_sphu = 0, tm_mixr = 0,
  tm_dswf = 0, tm_terr = 0)
}
\arguments{
\item{tratio}{The advection stability ratio. By default, this is set to
\code{0.75}.}

\item{initd}{The initial distribution. Set to \code{0} by default.}

\item{kpuff}{The horizontal puff dispersion growth. Setting to \code{0} results in
linear growth whereas option \code{1} uses an empirical growth scheme. By
default, this is set to \code{0} (linear growth).}

\item{khmax}{The maximum duration (in hours) for a particle or trajectory.
This is set to the absolute maximum by default, which is \code{9999} hours.}

\item{kmixd}{The methodology for modeling the mixed layer depth. There are
three options: \code{0} for using the input (the default), \code{1} for using
temperature, and \code{2} for using the TKE.}

\item{kmix0}{The minimum mixing depth in meters. This is \code{250} by default.}

\item{kzmix}{How to perform vertical mixing adjustments. There are three
options: \code{0} for not doing this at all (the default), \code{1} uses the PBL
average, and \code{3} scales with \emph{TVMIX}.}

\item{kdef}{The modeling method for horizontal turbulence. We can use the
vertical with option \code{0} (the default), or, use the deformation method with
option \code{1}.}

\item{kbls}{How to derive the boundary layer stability. Two options are
available for this. We can use fluxes (option \code{1}, default) or use wind
temperature (option \code{2}).}

\item{kblt}{The boundary layer turbulence parameterization to use. This can
either be Beljaars (option \code{1}), Kanthar (option \code{2}, default), or TKE
(option \code{3}).}

\item{conage}{Particle to- or from-puff conversion at \emph{CONAGE}. In units of
hours, with a default of \code{48}.}

\item{numpar}{The number of puffs or particles released per cycle. By default
this is set to \code{2500}.}

\item{qcycle}{An optional cycling of emissions, in units of hours.}

\item{efile}{An absolute path to an optional temporal emissions file.}

\item{tkerd}{The unstable turbulent kinetic energy ratio. This is set to
\code{0.18}.}

\item{tkern}{The stable turbulent kinetic energy ratio. This is set to
\code{0.18}.}

\item{ninit}{How to do particle initialization; with \code{0} no particle
initialization is done, with \code{1} this is done once, options \code{2} and \code{3} are
the \emph{add} and \emph{replace} methods.}

\item{ndump}{Should the particles be dumped to a file and, if so, how often?
Using \code{0} results in no writing particles to file (the default), and any
non-zero value performs this writing once per number of hours specified.}

\item{ncycl}{The \emph{PARDUMP} output cycle time.}

\item{pinpf}{The particle input file name (default is \code{"PARINIT"}). This is
useful for initialization or boundary conditions.}

\item{poutf}{The particle output  file name (default is \code{"PARDUMP"}).}

\item{mgmin}{The minimum meteorological subgrid size. The default value is
\code{10}.}

\item{kmsl}{The starting height reference. If it is to signify a distance
above ground level (AGL) then use option \code{0} (the default). If it is
instead relative to mean sea level, then option \code{1} should be used.}

\item{maxpar}{The maximum number of particles to be carried in simulation. By
default this is \code{10000}.}

\item{cpack}{The binary concentration packing. Here are the options: \code{0} for
none, \code{1} for nonzero, \code{2} for points, and \code{3} for polar. Option \code{1} is the
default option.}

\item{cmass}{Informs grid computation. The two options are to compute grid
concentrations (with \code{0}, the default) or to compute grid mass (option
\code{1}).}

\item{dxf, dyf}{The horizontal x- and y-grid adjustment factors for an
ensemble. By default, these are both set to \code{1.0}.}

\item{dzf}{The vertical factor for an ensemble. This is \code{0.01} by default.}

\item{ichem}{The chemistry conversion module to employ. Option \code{0} does no
chemistry (the default). With option \code{1} we use a matrix method, option \code{2}
does conversion, and option \code{3} works on dust.}

\item{maxdim}{The maximum number of pollutants to carry on one particle. By
default, this is \code{1}.}

\item{kspl}{The standard splitting interval in units of hours. By default,
this is \code{1} hour.}

\item{krnd}{The enhanced merge interval in hours. By default this is \code{6}
hours.}

\item{frhs}{The standard horizontal puff rounding fraction for the merge
process. This is \code{1.0} by default.}

\item{frvs}{The vertical puff rounding fraction, which is \code{0.01} by default.}

\item{frts}{The temporal puff rounding fraction, which is \code{0.10} by default.}

\item{frhmax}{The maximum value for the horizontal rounding parameter. This
is \code{3.0} by default.}

\item{splitf}{The automatic size adjustment factor for horizontal splitting.
By default, the value is \code{1.0}.}

\item{tm_pres, tm_tpot, tm_tamb, tm_rain, tm_mixd, tm_relh, tm_sphu, tm_mixr, tm_dswf, tm_terr}{Options to include meteorology along trajectory points. These are the
pressure variable marker flag (\code{tm_pres}), the potential temperature
(\code{tm_tpot}), the ambient temperature (\code{tm_tamb}), the rainfall rate
(\code{tm_rain}), the mixed layer depth (\code{tm_mixd}), the relative humidity
(\code{tm_relh}), the specific humidity (\code{tm_sphu}), the mixing rate
(\code{tm_mixr}), the downward short-wave flux (\code{tm_dswf}), and the terrain
height (\code{tm_terr}). Setting any of these to \code{0} disables output, whereas
\code{1} will enable output of these data points. By default, all are set to
\code{0}.}
}
\description{
Create a configuration list for a trajectory or dispersion model
}
