\name{Edit information in SplitR configuration files}
\alias{SplitR.info.edit}
\title{
Edit information in SplitR configuration files.
}
\description{
Helper function that automatically appends or removes information to and from the .SplitR configuration file that resides in the /SplitR folder. This function is used in conjunction with the 'SplitR.init' function, which initializes folders and information files.
}
\usage{
SplitR.info.edit(mode = "add",
                 heading,
                 string,
                 line_number)
}
\author{
Richard Iannone
}
\examples{
}
