% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajectory_list}
\alias{trajectory_list}
\title{List HYSPLIT trajectory output archive files or folders}
\usage{
trajectory_list(output_folder)
}
\arguments{
\item{output_folder}{the absolute path for the trajectory archive files (UNIX) or folders (Windows) is to be provided.}
}
\value{
data frame with information on trajectory model output data archives
}
\description{
The function lists the HYSPLIT trajectory output files that reside in a specified directory.
}

