% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_plot.R
\name{trajectory_plot}
\alias{trajectory_plot}
\title{Plot HYSPLIT trajectory model output onto a map}
\usage{
trajectory_plot(traj_df)
}
\arguments{
\item{traj_df}{a trajectory data frame.}
}
\description{
The function plots modeled wind
trajectories onto a map with information provided
at every hour of air transport.
}

