% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_read.R
\name{trajectory_read}
\alias{trajectory_read}
\title{Read HYSPLIT trajectory output files into a data
frame}
\usage{
trajectory_read(output_folder, year = NULL, start_height_m_AGL = NULL)
}
\arguments{
\item{output_folder}{the absolute path of the
directory containing the trajectory endpoints files
is to be provided.}

\item{year}{providing a year will filter the list
of trajectory endpoints files to be read into a data
frame. This is usually only effective for filtering
those trajectory endpoints files generated by SplitR
using the \code{hysplit_trajectory} function since
such files have a specified naming convention.}

\item{start_height_m_AGL}{providing a starting
height will filter the list of trajectory endpoints
files to be read into a data frame. This argument
can be used in conjunction with the \code{year}
argument in order to filter a directory of endpoints
files by year and by starting height. As with the
\code{year} argument, this option is usually only
effective for filtering those trajectory endpoints
files generated by SplitR.}
}
\value{
a data frame with several wind trajectory
parameters
}
\description{
The function takes HYSPLIT trajectory
output files in a specified output directory and
processes all files into a data frame object.
}
\examples{
\dontrun{
# Process all trajectory output files in the
# specified output directory
trajectory_df <-
  trajectory_read(
    output_folder = "traj--2015-06-16--23-58-44")

# Process trajectory output files in the same
# folder but only those files from 2002 with a
# starting height of 500 m
trajectory_df <-
  trajectory_read(
    output_folder = "traj--2015-06-16--23-58-44",
    year = 2002,
    start_height_m_AGL = 500)
}
}
