% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid.R
\name{add_grid}
\alias{add_grid}
\title{Add lat/lon grid to a model}
\usage{
add_grid(model, lat = 49.263, lon = -123.25, grid_ref = "center",
  n_s_dist = 5, w_e_dist = 5, n_s_division = 0.5, w_e_division = 0.5,
  dist_units = "degrees")
}
\arguments{
\item{model}{a SplitR modeling object}
}
\description{
Create and add a grid of latitude and
longitude points to a model object
}

