% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_params.R
\name{add_params}
\alias{add_params}
\title{Add model parameters}
\usage{
add_params(model, lat = NULL, lon = NULL, height = NULL,
  duration = NULL, days = NULL, start_day = NULL,
  start_hour = NULL, daily_hours = NULL, direction = NULL,
  met_type = NULL, vert_motion = NULL, model_height = NULL,
  traj_name = NULL, exec_dir = NULL, met_dir = NULL,
  binary_path = NULL)
}
\arguments{
\item{model}{A SplitR modeling object.}

\item{lat}{The starting latitude and longitude (in decimal degrees) for
each model run.}

\item{lon}{The starting latitude and longitude (in decimal degrees) for
each model run.}

\item{height}{The starting height (in meters above ground level) for each
model run.}

\item{duration}{The duration of each model run (either forward or backward)
in hours.}

\item{days}{A vector of days that the model will run. This is combined with
the \code{daily_hours} to produce a series of date-times.}

\item{daily_hours}{Should consist of a single daily hour as an integer hour
(from \code{0} to \code{23}), or, a vector of several daily hours represented as
integers.}

\item{direction}{An option to select whether to conduct the model in the
\code{"forward"} (default) or \code{"backward"} directions.}

\item{met_type}{An option to select meteorological data files. The options
are \code{"reanalysis"} (NCAR/NCEP global reanalysis data, the default),
\code{"gdas1"} (Global Data Assimilation System 1-degree resolution data), and
\code{"narr"} (North American Regional Reanalysis).}

\item{vert_motion}{A numbered option to select the method used to simulation
vertical motion. The methods are: \code{0} (input model data), \code{1} (isobaric),
\code{2} (isentropic), \code{3} (constant density), \code{4} (isosigma), \code{5} (from
divergence), \code{6} (remap MSL to AGL), \code{7} (average data), and \code{8} (damped
magnitude).}

\item{model_height}{The upper limit of the model domain in meters.}

\item{traj_name}{An optional, descriptive name for the output file
collection.}

\item{exec_dir}{An optional file path for the working directory of the model
input and output files.}

\item{met_dir}{An optional file path for storage and access of meteorological
data files.}

\item{binary_path}{An optional path to a HYSPLIT trajectory model binary.
When not specified, the model binary will be chosen from several available
in the package (based on the user's platform).}
}
\description{
Add modelling parameters to a model object.
}
