% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_params.R
\name{add_params}
\alias{add_params}
\title{Add model parameters}
\usage{
add_params(model, lat = NULL, lon = NULL, height = NULL,
  duration = NULL, run_period = NULL, start_day = NULL,
  start_hour = NULL, daily_hours = NULL, direction = NULL,
  met_type = NULL, vert_motion = NULL, model_height = NULL,
  traj_name = NULL, exec_dir = NULL, met_dir = NULL, binary_path = NULL)
}
\arguments{
\item{model}{a SplitR modeling object.}

\item{lat}{the starting latitude (in decimal 
degrees) for the model run(s).}

\item{lon}{the starting longitude (in decimal 
degrees) for the model run(s).}

\item{height}{the starting height (in meters above 
ground level) for the model run(s).}

\item{duration}{the duration of one or several
\code{backward}) in hours.}

\item{run_period}{the extended period (i.e., days,
years) when the model will initialize and run. This
can take the form of a single-length vector for a
day (\code{"YYYY-MM-DD"}) or year (\code{YYYY}), or,
a vector of length 2 to specify the range of days or
years.}

\item{start_day}{is specific to a dispersion model
and it refers to the day that the model will 
initialize and run. This should take the form of a
single-length vector for a day (\code{"YYYY-MM-DD"}).}

\item{start_hour}{is specific to a dispersion model 
and it should indicate which hour in the
\code{start_day} that modeling will begin. Provide
this as an integer (from \code{0} to \code{23}).}

\item{daily_hours}{are specific to a trajectory
model and should consist of a single daily hour as
an integer (from \code{0} to \code{23}), or, a
vector of several daily hours represented as
integers.}

\item{direction}{an option to select whether to
conduct the model in the \code{forward} or 
\code{backward} directions.}

\item{met_type}{an option to select meteorological
data files. The options are \code{gdas1} (Global
Data Assimilation System 1-degree resolution data), 
\code{reanalysis} (NCAR/NCEP global reanalysis
data), and \code{narr} (North American Regional 
Reanalysis).}

\item{vert_motion}{a numbered option to
select the method used to simulation vertical
motion. The methods are: \code{0} (input model
data), \code{1} (isobaric), \code{2} (isentropic),
\code{3} (constant density), \code{4} (isosigma),
\code{5} (from divergence), \code{6} (remap MSL to
AGL), \code{7} (average data), and \code{8} (damped
magnitude).}

\item{model_height}{the upper limit of the model
domain in meters.}

\item{traj_name}{an optional, descriptive name for
the output file collection.}

\item{exec_dir}{an optional file path for the
working directory of the model input and output
files.}

\item{met_dir}{an optional file path for storage and
access of meteorological data files.}

\item{binary_path}{an optional path to a HYSPLIT
trajectory model binary. When not specified, the
model binary will be chosen from several available
in the package (based on the user's platform).}
}
\description{
Add modelling parameters to a model
object
}
