% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create a grid of sites}
\usage{
create_grid(lat = 49.263, lon = -123.25, grid_ref = "center",
  n_s_dist = 5, w_e_dist = 5, n_s_division = 0.5, w_e_division = 0.5,
  dist_units = "degrees")
}
\arguments{
\item{lat}{a latitude value in decimal degrees
for the point of reference on the grid.}

\item{lon}{a longitude value in decimal degrees
for the point of reference on the grid.}

\item{grid_ref}{the grid reference point. The
default value is \code{center}.}

\item{n_s_dist}{the north-to-south distance of the
grid.}

\item{w_e_dist}{the west-to-east distance of the
grid.}

\item{n_s_division}{the north-to-south distances of
the those divisions in the grid.}

\item{w_e_division}{the west-to-east distances of
the those divisions in the grid.}

\item{dist_units}{the distance units used with
values supplied to  \code{n_s_dist},
\code{w_e_dist}, \code{n_s_division}, and
\code{w_e_division}. The default is \code{degrees}.}
}
\description{
Flexibly create a grid of sites for
ensemble trajectory model runs.
}

