% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trajectory_model.R
\name{create_trajectory_model}
\alias{create_trajectory_model}
\title{Create a trajectory model}
\usage{
create_trajectory_model(name = NULL)
}
\arguments{
\item{name}{An optional name for the trajectory model object.}
}
\description{
Create a trajectory model object to begin a modeling pipeline.
}
