\name{Trajectory model demonstration}
\alias{demo.trajectory}
\title{
Run an interactive demonstration for a trajectory model
}
\description{
Demonstrates the steps required to set up a project and get trajectory models running. For best viewing, increase the size of the console window to display at least 25 lines of text.
}
\usage{
demo.trajectory()
}
\author{
Richard Iannone
}
\examples{
# Run the trajectory demo
demo.trajectory()
}
