% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion.preset.add}
\alias{dispersion.preset.add}
\title{Add dispersion preset information}
\usage{
dispersion.preset.add(type, interactive = TRUE, species_name = NULL,
  particle_pdiam = NULL, particle_density = NULL,
  particle_shape_factor = NULL, ddep_velocity = NULL, ddep_MW = NULL,
  ddep_A_ratio = NULL, ddep_D_ratio = NULL, ddep_Henrys_Law_coeff = NULL,
  wdep_Henrys_Law_coeff = NULL, wdep_in_cloud_dep = NULL,
  wdep_below_cloud_dep = NULL, rad_decay = NULL,
  pollutant_resuspension_factor = NULL, emissions_name = NULL,
  emissions_rate = NULL, emissions_duration = NULL,
  emissions_start_time = NULL, grid_name = NULL, grid_center = NULL,
  grid_spacing = NULL, grid_span = NULL, grid_folder = NULL,
  grid_filename = NULL, grid_number_vertical = NULL, grid_heights = NULL,
  grid_start_time = NULL, grid_end_time = NULL,
  sampling_interval_type_rate = NULL,
  path_wd = "~/Documents/SplitR/Working/")
}
\arguments{
\item{type}{}

\item{interactive}{}

\item{species_name}{}

\item{particle_pdiam}{}

\item{particle_density}{}

\item{particle_shape_factor}{}

\item{ddep_velocity}{}

\item{ddep_MW}{}

\item{ddep_A_ratio}{}

\item{ddep_D_ratio}{}

\item{ddep_Henrys_Law_coeff}{}

\item{wdep_Henrys_Law_coeff}{}

\item{wdep_in_cloud_dep}{}

\item{wdep_below_cloud_dep}{}

\item{rad_decay}{}

\item{pollutant_resuspension_factor}{}

\item{emissions_name}{}

\item{emissions_rate}{}

\item{emissions_duration}{}

\item{emissions_start_time}{}

\item{grid_name}{}

\item{grid_center}{}

\item{grid_spacing}{}

\item{grid_span}{}

\item{grid_folder}{}

\item{grid_filename}{}

\item{grid_number_vertical}{}

\item{grid_heights}{}

\item{grid_start_time}{}

\item{grid_end_time}{}

\item{sampling_interval_type_rate}{}

\item{path_wd}{}
}
\description{
Add dispersion preset information
}
\examples{
# Interactively add a preset for a 'emissions' input
dispersion.preset.add("emissions")

# Add a default preset for a 'species' input
dispersion.preset.add("species", interactive = FALSE,
                      species_name = 'default')

# Add a preset for a 'grids' input
dispersion.preset.add("grids", interactive = FALSE,
                      grid_name = "MainGrid",
                      grid_center = "36.5 -122.4",
                      grid_spacing = "0.05 0.05",
                      grid_span = "20 20",
                      grid_number_vertical = 5,
                      grid_heights = "0 200 400 800 1200",
                      grid_start_time = "12 02 01 12 00",
                      grid_end_time = "12 02 08 12 00")
}

