% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion.preset.delete}
\alias{dispersion.preset.delete}
\title{Remove dispersion preset information}
\usage{
dispersion.preset.delete(read = NULL, numbers = NULL, interactive = TRUE,
  path_wd = "~/Documents/SplitR/Working/")
}
\arguments{
\item{read}{}

\item{numbers}{}

\item{interactive}{}

\item{path_wd}{}
}
\description{
Remove dispersion preset information
}
\examples{
# Delete an 'emissions' preset
dispersion.preset.delete("emissions")

# Delete a 'grids' preset
dispersion.preset.delete("grids")

# Delete a 'species' preset
dispersion.preset.delete("species")
}

