% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion.preset.get}
\alias{dispersion.preset.get}
\title{Get requested preset information}
\usage{
dispersion.preset.get(read, numbers, path_wd = "~/Documents/SplitR/Working/")
}
\arguments{
\item{read}{the type of preset to get. The three preset types are "emissions", "grids", and "species". Only a single type can be specified.}

\item{numbers}{the values corresponding to the presets of a specified type (provided as the 'read' character argument) to be retrieved. The argument is to be provided as a single numeric value or numeric vector. The numbers correspond to the order in which the presets appear in each of the 'emissions', 'grids', or 'species' plaintext files.}

\item{path_wd}{}
}
\description{
Get requested preset information
}
\examples{
# The hysplit.dispersion() function obtains presets for
# "emissions", "grids", and "species" while assembling
# a CONTROL file
dispersion.preset.get(read = "emissions", numbers = emissions)
dispersion.preset.get(read = "grids", numbers = grids)
dispersion.preset.get(read = "species", numbers = species)
}

