\name{Get requested preset information}
\alias{dispersion.preset.get}
\title{
Obtain requested presets in a format suitable for inclusion in a CONTROL file
}
\description{

}
\usage{
dispersion.preset.get(read, numbers)
}
\arguments{
  \item{read}{
the type of preset to get. The three preset types are "emissions", "grids", and "species". Only a single type can be specified.
}

}
\author{
Richard Iannone
}
\examples{

}
