% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion.read}
\alias{dispersion.read}
\title{Read HYSPLIT dispersion output files}
\usage{
dispersion.read(path_output_files)
}
\arguments{
\item{path_output_files}{the absolute path of the dispersion output files is to be provided.}
}
\value{
data frame with hourly particle positions
}
\description{
The function takes HYSPLIT dispersion output files in a specified output directory and processes all files into a data frame object.
}
\examples{
# Process all dispersion output files in the specified output directory
dispersion.df <- dispersion.read(path_output_files =
                                 "~/hysplit/dispersion/")
}

