% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_plot.R
\name{dispersion_plot}
\alias{dispersion_plot}
\title{Plot HYSPLIT dispersion model output onto a map}
\usage{
dispersion_plot(x, color_scheme = "cycle_hues")
}
\arguments{
\item{x}{Either a dispersion data frame, typically created from use of the
\code{hysplit_dispersion}, or a dispersion model object that contains
output data (i.e., after executing model runs via the \code{run_model}
function).}

\item{color_scheme}{Defines the appearance of multiple trajectories in a
single plot. Current options are \code{cycle_hues} (the default), and
\code{increasingly_gray}.}
}
\description{
The function plots modeled dispersion particles onto an interactive map.
}
