% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_read.R
\name{dispersion_read}
\alias{dispersion_read}
\title{Read HYSPLIT dispersion output files}
\usage{
dispersion_read(archive_folder)
}
\arguments{
\item{archive_folder}{the absolute path of the
directory containing dispersion output files is 
provided here.}
}
\value{
data frame with hourly particle positions
}
\description{
The function takes HYSPLIT dispersion
output files in a specified output directory and
processes all files into a data frame object.
}
\examples{
\dontrun{
# Process all dispersion output files in the 
# specified output directory
dispersion.df <- 
  dispersion_read(
    path_output_files =  "~/hysplit/dispersion/")
}
}
