% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.edas40}
\alias{get.met.edas40}
\title{Get EDAS40 meteorology data files}
\usage{
get.met.edas40(files = NULL, years = NULL, months = NULL, path_met_files)
}
\arguments{
\item{files}{a vector list of exact filenames for the EDAS40 files.}

\item{years}{a vector list of years for which EDAS40 files are to be obtained via FTP.}

\item{months}{a vector list of months for which EDAS40 files are to be obtained via FTP.}

\item{path_met_files}{a full path should be provided for the location of the meteorological data files; downloaded files will be saved in this location.}
}
\description{
This function downloads EDAS40 meteorology data files from the NOAA FTP server and saves them to a specified folder. Files can be downloaded either by specifying a list of filenames (in the form of "RP[YYYY][MM].gbl") or through bulk download of a year of files.
}
\examples{
\dontrun{
# Download a year of EDAS40 data files
get.met.edas40(years = 2012)
}
}

