% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.edas40}
\alias{get.met.edas40}
\title{Get EDAS40 meteorology data files}
\usage{
get.met.edas40(files = NULL, years = NULL, months = NULL, path_met_files)
}
\arguments{
\item{files}{}

\item{years}{}

\item{months}{}

\item{path_met_files}{}
}
\description{
This function downloads EDAS40 meteorology data files from the NOAA FTP server and saves them to a specified folder. Files can be downloaded either by specifying a list of filenames (in the form of "RP[YYYY][MM].gbl") or through bulk download of a year of files.
}
\examples{
# Download a year of EDAS40 data files
get.met.edas40(years = 2012)
}

