% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.gdas0p5}
\alias{get.met.gdas0p5}
\title{Get GDAS0.5 meteorology data files}
\usage{
get.met.gdas0p5(files = NULL, years = NULL, months = NULL, days = NULL,
  path_met_files = "~/Documents/SplitR/Met/")
}
\arguments{
\item{files}{}

\item{years}{}

\item{months}{}

\item{days}{}

\item{path_met_files}{}
}
\description{
Get GDAS0.5 meteorology data files
}

