% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.gdas1}
\alias{get.met.gdas1}
\title{Get GDAS1 meteorology data files}
\usage{
get.met.gdas1(files = NULL, years = NULL, months = NULL,
  path_met_files = "~/Documents/SplitR/Met/")
}
\arguments{
\item{files}{}

\item{years}{}

\item{months}{}

\item{path_met_files}{}
}
\description{
Get GDAS1 meteorology data files
}

