% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.met.nam12}
\alias{get.met.nam12}
\title{Get NAM12 meteorology data files}
\usage{
get.met.nam12(files = NULL, years = NULL, months = NULL,
  path_met_files = "~/Documents/SplitR/Met/")
}
\arguments{
\item{files}{}

\item{years}{}

\item{months}{}

\item{path_met_files}{}
}
\description{
This function downloads NAM12 meteorology data files from the NOAA FTP server and saves them to a specified folder. Files can be downloaded either by specifying a list of filenames (in the form of "RP[YYYY][MM].gbl") or through bulk download of a year of files.
}

