% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_forecast_nam.R
\name{get_met_forecast_nam}
\alias{get_met_forecast_nam}
\title{Read NAM meteorological forecast data files}
\usage{
get_met_forecast_nam(path_met_files)
}
\arguments{
\item{path_met_files}{A full directory path to which the meteorological data
files will be saved.}
}
\description{
This function downloads NAM meteorology forecast data files from the NOAA FTP
server and saves them to a specified folder.
}
