% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_narr.R
\name{get_met_narr}
\alias{get_met_narr}
\title{Get NARR meteorology data files}
\usage{
get_met_narr(days, duration, direction, path_met_files)
}
\arguments{
\item{days}{A vector of days that the model will run. This is combined with
the \code{daily_hours} to produce a series of date-times.}

\item{duration}{The duration of each model run (whether it is in the forward
direction or running backwards) in hours.}

\item{direction}{An option to select whether to conduct the model in the
\code{"forward"} (default) or \code{"backward"} directions.}

\item{path_met_files}{A full directory path to which the meteorological data
files will be saved.}
}
\description{
This function downloads NARR meteorology data files from the NOAA FTP server
and saves them to a specified folder. Files can be downloaded by specifying a
list of filenames (in the form of \code{"NARR{YYYY}{MM}"}).
}
