% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met_reanalysis.R
\name{get_met_reanalysis}
\alias{get_met_reanalysis}
\title{Get reanalysis meteorology data files}
\usage{
get_met_reanalysis(files = NULL, years = NULL, path_met_files)
}
\arguments{
\item{files}{a vector list of exact filenames for
the reanalysis files.}

\item{years}{a vector list of years for which
reanalysis files are to be obtained via FTP.}

\item{path_met_files}{a full path should be provided
for the location of the meteorological data files;
downloaded files will be saved in this location.}
}
\description{
This function downloads reanalysis
meteorology data files from the NOAA FTP server and
saves them to a specified folder. Files can be
downloaded either by specifying a list of filenames
(in the form of \code{'RP[YYYY][MM].gbl'}) or
through bulk download of a year of files.
}
