% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hysplit.dispersion.config}
\alias{hysplit.dispersion.config}
\title{Create config file for dispersion runs}
\usage{
hysplit.dispersion.config(path_wd = "~/Documents/SplitR/Working/")
}
\arguments{
\item{path_wd}{}
}
\description{
Create config file for dispersion runs
}
\examples{
# Create a default SETUP.CFG file for dispersion runs
hysplit.dispersion.config(path_wd = "C://hysplit//working//")
}

