\name{Create configuration file for dispersion runs}
\alias{hysplit.dispersion.config}
\title{
Create or modify the configuration file used for HYSPLIT dispersion models
}
\description{

}
\usage{
hysplit.dispersion.config(create_default_config = TRUE,
                          interactive_mode = FALSE)
}
\arguments{

}
\details{
}
\value{
}
\references{
}
\author{
Richard Iannone
}
\examples{

}
