% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hysplit.dispersion.config}
\alias{hysplit.dispersion.config}
\title{Create config file for dispersion runs}
\usage{
hysplit.dispersion.config(path_wd = "~/Documents/SplitR/Working/")
}
\arguments{
\item{path_wd}{}
}
\description{
This function creates a default SETUP.CFG configuration file and places it in the user-specified 'working' folder. This file must be located in the working directory before any HYSPLIT dispersion models can be run.
}
\examples{
# Create a default SETUP.CFG file for dispersion runs
hysplit.dispersion.config(path_wd = "C://hysplit//working//")
}

