% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_dispersion.R
\name{hysplit_dispersion}
\alias{hysplit_dispersion}
\title{Conduct HYSPLIT dispersion runs}
\usage{
hysplit_dispersion(lat = 49.263, lon = -123.25, height = 50,
  duration = 24, run_type = "day", run_day = "2013-02-01",
  run_range = c("2013-02-01", "2013-02-10"), run_years = "2013",
  daily_hours = "00", backward_running = FALSE, met_type = "reanalysis",
  vert_motion = 0, model_height = 20000, particle_num = 2500,
  particle_max = 10000, emissions = 1, species = 1, grids = 1,
  return_disp_df = TRUE, write_disp_CSV = TRUE, disp_name = NULL)
}
\arguments{
\item{lat}{the starting latitude (in decimal
degrees) for the model run(s).}

\item{lon}{the starting longitude (in decimal
degrees) for the model run(s).}

\item{height}{the starting height (in meters above
ground level) for the model run(s).}

\item{duration}{the duration of each
model run (either forward or backward) in hours.}

\item{run_type}{used to select whether models should
be run for a single day (\code{day}), for one or
more years (\code{years}), or within a specified
date range (\code{range}).}

\item{run_day}{used when \code{run_type} of
\code{day} is selected. The date format should be
provided here as "YYYY-MM-DD".}

\item{run_range}{used when 'run_type' of 'range' is
selected. The date format should be provided here
as \code{c("YYYY-MM-DD", "YYYY-MM-DD")}.}

\item{run_years}{used when 'run_type' of 'years' is
selected. The format should either be a single year
("YYYY") or a range of years ("YYYY-YYYY").}

\item{daily_hours}{should consist of a
single daily hour in the format "HH", or, several
daily hours in the format \code{c("HH", "HH", "HH", ...)}.}

\item{backward_running}{an option to select whether
the dispersion runs should be running forward in
time (the default) or in a backward running state.}

\item{met_type}{an option to select meteorological
data files. The options are \code{gdas1} (Global Data
Assimilation System 1-degree resolution data) and
\code{reanalysis} (NCAR/NCEP global reanalysis data).}

\item{vert_motion}{a numbered option to
select the method used to simulation vertical
motion. The methods are: (0) input model data,
(1) isobaric, (2) isentropic, (3) constant density,
(4) isosigma, (5) from divergence, (6) remap MSL to
AGL, (7) average data, and (8) damped magnitude.}

\item{model_height}{the upper limit of the
model domain in meters.}

\item{particle_num}{the number of particles
released by source during each release cycle.}

\item{particle_max}{the number of particles
released by a source during a model run.}

\item{emissions}{the numbers corresponding to the
stored emissions presets. These presets are
specified using the function
\code{hysplit_dispersion_define("emissions")}.}

\item{species}{the numbers corresponding to the
stored species presets. These presets are specified
using the function
\code{hysplit_dispersion_define("species")}.}

\item{grids}{the numbers corresponding to the
stored grid presets. These presets are specified
using the function
\code{hysplit_dispersion_define("grids")}.}

\item{return_disp_df}{an option to return a data
frame with dispersion data.}

\item{write_disp_CSV}{an option to write disperison
data to a CSV file.}

\item{disp_name}{an optional, descriptive name for
the output file collection.}
}
\description{
The function executes single/multiple
forward or backward HYSPLIT dispersion runs using
specified meteorological datasets.
}
\examples{
\dontrun{
# Perform a dispersion run lasting 12 hours over
# two consecutive days, both starting at midnight;
# grid presets 1 and 2 will be used as sampling
# grids (presets for species and emissions are set
# to the first, which is the default)
hysplit_dispersion(
  disp_name = "example",
  return_disp_df = FALSE,
  write_disp_CSV = TRUE,
  lat = 49.289328,
  lon = -123.117665,
  height = 15,
  duration = 12,
  backward_running = FALSE,
  emissions = 1,
  species = 1,
  grids = c(1,2),
  run_type = "range",
  run_range = c("2012-02-01", "2012-02-03"),
  daily_hours = "00")
}
}

