% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_dispersion_plot.R
\name{hysplit_dispersion_plot}
\alias{hysplit_dispersion_plot}
\title{Plot HYSPLIT dispersion model output onto maps}
\usage{
hysplit_dispersion_plot(hours = "all", dispersion_df = NULL,
  disp_folder_path = NULL, map_type = "stamen", map_output_name = NULL)
}
\arguments{
\item{hours}{specification of hours of dispersion
data to plot.}

\item{dispersion_df}{optionally specify a data frame
with dispersion data as an input.}

\item{disp_folder_path}{absolute path of the folder
containing a dispersion data frame.}

\item{map_type}{selection provider of base maps for
plotting. Choices are 'osm' (Open Street Map) and
'stamen' (Stamen Maps).}

\item{map_output_name}{a partial identifier
prepended to the output map files.}
}
\description{
The function plots hourly outputs of
dispersed particles onto maps.
}
\examples{
\dontrun{
# Make a set of hourly plots from a dispersion
# data frame
hysplit_dispersion_plot(
  hours = "all",
  dispersion_df = disp.df,
  map_output_name = "new map")
}
}

