% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_trajectory.R
\name{hysplit_trajectory}
\alias{hysplit_trajectory}
\title{Conduct HYSPLIT trajectory runs}
\usage{
hysplit_trajectory(traj_name = NULL, return_traj_df = TRUE,
  start_lat_deg = 49.263405, start_long_deg = -123.250054,
  start_height_m_AGL = 50, simulation_duration_h = 24,
  backtrajectory = FALSE, met_type = "reanalysis",
  vertical_motion_option = 0, top_of_model_domain_m = 20000,
  run_type = "day", run_day = "2015-07-01", run_range = NULL,
  run_years = NULL, daily_hours_to_start = 0,
  return_met_along_traj = FALSE)
}
\arguments{
\item{traj_name}{an optional, descriptive name for
the output file collection.}

\item{return_traj_df}{an option to return a data
frame with trajectory data.}

\item{start_lat_deg}{the starting latitude (in
decimal degrees) for the model run(s).}

\item{start_long_deg}{the starting longitude (in
decimal degrees) for the model run(s).}

\item{start_height_m_AGL}{the starting height (in
meters above ground level) for the model run(s).}

\item{simulation_duration_h}{the duration of each
model run (either forward or backward) in hours.}

\item{backtrajectory}{an option to select whether to
conduct forward trajectory model runs or to generate
backtrajectories.}

\item{met_type}{an option to select meteorological
data files. The options are \code{gdas1} (Global
Data Assimilation System 1-degree resolution data), 
\code{reanalysis} (NCAR/NCEP global reanalysis
data), and \code{narr} (North American Regional 
Reanalysis).}

\item{vertical_motion_option}{a numbered option to
select the method used to simulation vertical
motion. The methods are: \code{0} (input model
data), \code{1} (isobaric), \code{2} (isentropic),
\code{3} (constant density), \code{4} (isosigma),
\code{5} (from divergence), \code{6} (remap MSL to
AGL), \code{7} (average data), and \code{8} (damped
magnitude).}

\item{top_of_model_domain_m}{the upper limit of the
model domain in meters.}

\item{run_type}{used to select whether models should
be run for a single day (\code{day}), for one or
more years (\code{years}), or within a specified
date range (\code{range}).}

\item{run_day}{used when the \code{run_type} of
\code{day} is selected. The date format should be
provided here in the format \code{YYYY-MM-DD}.}

\item{run_range}{used when the \code{run_type} of
\code{range} is selected. The date format should be
provided here as a vector of length 2 in the form:
 \code{c("YYYY-MM-DD", "YYYY-MM-DD")}.}

\item{run_years}{used when the \code{run_type} of
\code{years} is selected. The format should either
be a single year (\code{YYYY}) or a range of years
(\code{YYYY-YYYY}).}

\item{daily_hours_to_start}{should consist of a
single daily hour in the format \code{HH}, or,
several daily hours as a vector in the format of
\code{"c("HH", "HH", ...)"}.}

\item{return_met_along_traj}{an option to report
additional meteorological data along each output
trajectory.}
}
\description{
The function executes single/multiple
forward or backward HYSPLIT trajectory runs using
specified meteorological datasets.
}
\examples{
\dontrun{
# Test with a run type of `years` with a forward
# trajectory using NCEP/NCAR reanalysis data
hysplit_trajectory(
  traj_name = "second",
  return_traj_df = FALSE,
  start_lat_deg = 50.108,
  start_long_deg = -122.942,
  start_height_m_AGL = 200.0,
  simulation_duration_h = 96,
  backtrajectory = FALSE,
  met_type = "reanalysis",
  vertical_motion_option = 0,
  top_of_model_domain_m = 20000,
  run_type = "years",
  run_years = "2004",
  daily_hours_to_start = c(3, 6, 9, 12, 15, 18, 21),
  return_met_along_traj = TRUE)
}
}

