% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_trajectory.R
\name{hysplit_trajectory}
\alias{hysplit_trajectory}
\title{Conduct HYSPLIT trajectory runs}
\usage{
hysplit_trajectory(lat = 49.263, lon = -123.25, height = 50,
  duration = 24, run_period = "2015-07-01", daily_hours = 0,
  direction = "forward", met_type = "reanalysis", vert_motion = 0,
  model_height = 20000, extended_met = FALSE, return_traj_df = TRUE,
  traj_name = NULL, exec_dir = NULL, met_dir = NULL,
  binary_path = NULL)
}
\arguments{
\item{lat, lon}{The starting latitude and longitude (in decimal degrees) for
each model run.}

\item{height}{The starting height (in meters above ground level) for each
model run.}

\item{duration}{The duration of each model run (either forward or backward)
in hours.}

\item{run_period}{The extended period (i.e., days, years) when the model will
initialize and run. This can take the form of a single-length vector for a
day (\code{"YYYY-MM-DD"}) or year (\code{"YYYY"}), or, a vector of length 2 to
specify the range of days or years.}

\item{daily_hours}{Should consist of a single daily hour as an integer hour
(from \code{0} to \code{23}), or, a vector of several daily hours represented as
integers.}

\item{direction}{An option to select whether to conduct the model in the
\code{"forward"} (default) or \code{"backward"} directions.}

\item{met_type}{An option to select meteorological data files. The options
are \code{"reanalysis"} (NCAR/NCEP global reanalysis data, the default),
\code{"gdas1"} (Global Data Assimilation System 1-degree resolution data),
\code{"narr"} (North American Regional Reanalysis), and \code{"hrrr"} (High
Resolution Rapid Refresh 3-km resolution data - CONUS only).}

\item{vert_motion}{A numbered option to select the method used to simulation
vertical motion. The methods are: \code{0} (input model data), \code{1} (isobaric),
\code{2} (isentropic), \code{3} (constant density), \code{4} (isosigma), \code{5} (from
divergence), \code{6} (remap MSL to AGL), \code{7} (average data), and \code{8} (damped
magnitude).}

\item{model_height}{The upper limit of the model domain in meters.}

\item{extended_met}{An option to report additional meteorological data along
each output trajectory.}

\item{return_traj_df}{An option to return a data frame with trajectory data.}

\item{traj_name}{An optional, descriptive name for the output file
collection.}

\item{exec_dir}{An optional file path for the working directory of the model
input and output files.}

\item{met_dir}{An optional file path for storage and access of meteorological
data files.}

\item{binary_path}{An optional path to a HYSPLIT trajectory model binary.
When not specified, the model binary will be chosen from several available
in the package (based on the user's platform).}
}
\description{
The function executes single/multiple forward or backward HYSPLIT trajectory
runs using specified meteorological datasets.
}
\examples{
\dontrun{
# Run a trajectory model 4 times a day throughout
# 2004 using NCEP/NCAR reanalysis data
trajectory <-
  hysplit_trajectory(
    lat = 50.108,
    lon = -122.942,
    height = 100,
    duration = 48,
    run_period = 2004,
    daily_hours = c(0, 6, 12, 18)
  )
}
}
