\name{Read HYSPLIT trajectory output files}
\alias{trajectory.read}
\title{
Read HYSPLIT trajectory output files into data frame for further analysis.
}
\description{
The function takes HYSPLIT trajectory output files in a specified output directory and processes all files into a data frame object. The resulting data frame is useful for plotting of trajectories and for conducting statistical analyses such as clustering.  
}
\usage{
trajectory.read(path_output_files,
                year = NULL,
                start_height_m_AGL = NULL)
}
\arguments{
  \item{path_output_files}{
a full path should be provided for a location that contains all relevant trajectory output files.
}
  \item{year}{
providing a year will filter the list of trajectory endpoints files to be read into a data frame. This is usually only effective for filtering those trajectory endpoints files generated by SplitR using the the 'hysplit.trajectory' function since such files have a specified naming convention.
}
  \item{start_height_m_AGL}{
providing a starting height will filter the list of trajectory endpoints files to be read into a data frame. This argument can be used in conjunction with the 'year' argument in order to filter a directory of endpoints files by year and by starting height. As with the 'year' argument, this option is usually only effective for filtering those trajectory endpoints files generated by SplitR.
}
}
\author{
Richard Iannone
}
\examples{
# Process all trajectory output files in the specified output directory
trajectory.df <- trajectory.read(path_output_files =
                                 "~/hysplit/trajectory")

# Process trajectory output files in the same folder but only those
# files from 2002 with a starting height of 500 m
trajectory.df <- trajectory.read(path_output_files =
                                 "~/hysplit/trajectory",
                                 year = 2002,
                                 start_height_m_AGL = 500)

}
