% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajectory.read}
\alias{trajectory.read}
\title{Read HYSPLIT trajectory output files}
\usage{
trajectory.read(archive_folder, year = NULL, start_height_m_AGL = NULL)
}
\arguments{
\item{archive_folder}{the absolute path and filename of the trajectory archive file (UNIX) or folder (Windows) is to be provided.}

\item{year}{providing a year will filter the list of trajectory endpoints files to be read into a data frame. This is usually only effective for filtering those trajectory endpoints files generated by SplitR using the the 'hysplit.trajectory' function since such files have a specified naming convention.}

\item{start_height_m_AGL}{providing a starting height will filter the list of trajectory endpoints files to be read into a data frame. This argument can be used in conjunction with the 'year' argument in order to filter a directory of endpoints files by year and by starting height. As with the 'year' argument, this option is usually only effective for filtering those trajectory endpoints files generated by SplitR.}
}
\value{
data frame with several wind trajectory parameters
}
\description{
The function takes HYSPLIT trajectory output files in a specified output directory and processes all files into a data frame object.
}
\examples{
# Process all trajectory output files in the specified output directory
trajectory.df <- trajectory.read(archive_folder =
                                 "~/hysplit/output/traj--2014-06-16--23-58-44.zip")

# Process trajectory output files in the same folder but only those
# files from 2002 with a starting height of 500 m
trajectory.df <- trajectory.read(archive_folder =
                                   "~/hysplit/output/traj--2014-06-16--23-58-44.zip",
                                 year = 2002,
                                 start_height_m_AGL = 500)
}

