% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_read.R
\name{trajectory_read}
\alias{trajectory_read}
\title{Read HYSPLIT trajectory output files into a data frame}
\usage{
trajectory_read(output_folder)
}
\arguments{
\item{output_folder}{The path of the directory containing the trajectory
endpoints files.}
}
\value{
A tibble with HYSPLIT trajectory data.
}
\description{
The function takes HYSPLIT trajectory output files in a specified output
directory and processes all files into a data frame object.
}
\examples{
\dontrun{
# Process all trajectory output files in the
# specified output directory
trajectory_df <-
  trajectory_read(
    output_folder = "traj--2015-06-16--23-58-44")
}
}
